# internal/docs/swagger.yaml
definitions:
  handler.appointmentResponse:
    properties:
      created_at:
        type: string
      doctor_id:
        type: integer
      doctor_name:
        type: string
      id:
        type: integer
      notes:
        type: string
      patient_id:
        type: integer
      patient_name:
        type: string
      reason:
        type: string
      scheduled_end:
        type: string
      scheduled_start:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  handler.completeAppointmentRequest:
    properties:
      notes:
        type: string
    type: object
  handler.createAppointmentRequest:
    properties:
      doctor_id:
        type: integer
      notes:
        type: string
      patient_id:
        type: integer
      reason:
        type: string
      scheduled_end:
        description: RFC3339 format
        type: string
      scheduled_start:
        description: RFC3339 format
        type: string
      type:
        description: in_person, video, phone
        type: string
    required:
    - doctor_id
    - patient_id
    - scheduled_end
    - scheduled_start
    type: object
...
  handler.paginatedAppointmentsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/handler.appointmentResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
    type: object
...
  handler.updateAppointmentRequest:
    properties:
      notes:
        type: string
      reason:
        type: string
      scheduled_end:
        description: RFC3339 format
        type: string
      scheduled_start:
        description: RFC3339 format
        type: string
      status:
        type: string
      type:
        description: in_person, video, phone
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /appointments:
    post:
      consumes:
      - application/json
      description: Create a new appointment for a patient with a doctor
      parameters:
      - description: Appointment Details
        in: body
        name: appointment
        required: true
        schema:
          $ref: '#/definitions/handler.createAppointmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Appointment created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new appointment
      tags:
      - appointments
  /appointments/{id}:
    get:
      consumes:
      - application/json
      description: Get appointment details by ID
      parameters:
      - description: Appointment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Appointment
          schema:
            $ref: '#/definitions/handler.appointmentResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get appointment by ID
      tags:
      - appointments
    put:
      consumes:
      - application/json
      description: Update an existing appointment
      parameters:
      - description: Appointment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Appointment Details
        in: body
        name: appointment
        required: true
        schema:
          $ref: '#/definitions/handler.updateAppointmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Appointment updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update appointment
      tags:
      - appointments
  /appointments/{id}/cancel:
    post:
      consumes:
      - application/json
      description: Cancel an existing appointment
      parameters:
      - description: Appointment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Appointment cancelled successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cancel appointment
      tags:
      - appointments
  /appointments/{id}/complete:
    post:
      consumes:
      - application/json
      description: Mark an appointment as completed
      parameters:
      - description: Appointment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Completion Details
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handler.completeAppointmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Appointment completed successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Complete appointment
      tags:
      - appointments
...
  /doctors/{doctor_id}/appointments:
    get:
      consumes:
      - application/json
      description: Get appointments for the specified doctor
      parameters:
      - description: Doctor ID
        in: path
        name: doctor_id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Doctor appointments
          schema:
            $ref: '#/definitions/handler.paginatedAppointmentsResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get doctor appointments
      tags:
      - appointments
      - doctors
  /doctors/{doctor_id}/schedule:
    get:
      consumes:
      - application/json
      description: Get doctor's schedule for a date range
      parameters:
      - description: Doctor ID
        in: path
        name: doctor_id
        required: true
        type: integer
      - description: Start date (RFC3339 format)
        in: query
        name: start_date
        type: string
      - description: End date (RFC3339 format)
        in: query
        name: end_date
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Doctor schedule
          schema:
            $ref: '#/definitions/handler.paginatedAppointmentsResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get doctor schedule
      tags:
      - appointments
      - doctors
...
  /patients/{patient_id}/appointments:
    get:
      consumes:
      - application/json
      description: Get appointments for the specified patient
      parameters:
      - description: Patient ID
        in: path
        name: patient_id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Patient appointments
          schema:
            $ref: '#/definitions/handler.paginatedAppointmentsResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get patient appointments
      tags:
      - appointments
      - patients
...
swagger: "2.0"